Soundsystem tweaks - two ignore vars for better soundproof manipulation SFX system will need to be adapted to new sound system. Key sound effects for locking/unlocking/trying to open door! Moved item carrying to player.cs Added CleanBlood event for sinks/showers to use

+++ lib/items.cs
@@ -10,14 +10,16 @@ function Player::addTool(%this, %data, %props)
 
 		if (!%data.customPickupMultiple && %this.tool[%i] == %data)
 		{
-			%props.delete();
+			if (isObject(%props))
+				%props.delete();
 			return -1;
 		}
 	}
 
 	if (%i == %maxTools)
 	{
-		%props.delete();
+		if (isObject(%props))
+			%props.delete();
 		return -1;
 	}
